backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
psycopg[binary]==3.2.12
sqlparse==0.5.3
typing_extensions==4.15.0
python-dotenv==1.0.1
dj-database-url==2.3.0
djangorestframework==3.16.1
